+++ src/card/sets/tests/runtime_support/nested_definitions.rs
@@ -356,6 +356,7 @@ pub(in super::super) fn assert_nested_definition_applied_effect(
         AppliedEffectDef::CannotBeCountered
         | AppliedEffectDef::DoesNotUntapDuringUntapStep
         | AppliedEffectDef::MayChooseNotToUntap
+        | AppliedEffectDef::CannotBlock
         | AppliedEffectDef::CannotBeEnchanted
         | AppliedEffectDef::CannotBecomeEnchanted
         | AppliedEffectDef::CannotChangeController